Episode 3: Converting to Vector3s and adding a rudimentary Rigidbody class

+++ Twitch/Engine/Engine.cpp
@@ -5,7 +5,7 @@
 int Engine::SCREEN_WIDTH = 1024;
 int Engine::SCREEN_HEIGHT = 768;
 GLFWwindow* Engine::window = NULL;
-double Engine::dt = 0;
+float Engine::dt = 0;
 
 Engine::Engine()
 {
@@ -61,14 +61,14 @@ bool Engine::Initialize(char* windowTitle)
 	glEnable(GL_BLEND);
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
-	lastTime = glfwGetTime();
+	lastTime = (float)glfwGetTime();
 
 	return true;
 }
 
 void Engine::Update()
 {
-	double now = glfwGetTime();
+	float now = (float)glfwGetTime();
 	dt = (now - lastTime);
 	lastTime = now;
 	glfwPollEvents();
@@ -85,7 +85,7 @@ void Engine::EndRender()
 	glfwSwapBuffers(window);
 }
 
-double Engine::GetDT()
+float Engine::GetDT()
 {
 	return dt;
 }
